feat(UIPointer): emit events when pointer enters/exits valid target
The UI Pointer now emits a `UIPointerElementEnter` event when the
UI Pointer beam collides with a valid UI Canvas or UI element.

When the UI Pointer beam is no longer touching any valid element then
the `UIPointerElementExit` event is emitted.

The event payload contains the controller id of the UI Pointer, the
current target the pointer is colliding with, and the previous target
that it was colliding with.

The scene `034_Controls_InteractingWithUnityUI` has been updated with
a test script that shows the event calls in the console and
demonstrates how to automatically toggle the pointer beam if the UI
Pointer is colliding with a valid canvas.