feat(Controller): add undefined button alias
It's now possible to set a button alias as `Undefined` which specifies
that the button alias will not be triggered by any button press.

This is useful if an action associated to a button requires to be
disabled as it can simply be done by setting the alias to `Undefined`.