fix(Interaction): prevent custom rigidbody gameobject instantiation
If a Custom Rigid Body object is provided on Interact Touch and it
exists within the controller hierarchy then it should not be
instantiated as this should only happen for prefabs (or the default
collider which is also a prefab).

This fix, ensures that if a custom rigid body is given and it exists
in the scene, then it won't be cloned and the original will be used.