fix(Interaction): prevent controller rigidbody from being destroyed
Previously, the controller rigidbody would get created in the
`OnEnable` call and deleted in the `OnDisable` call. However,
deleting the rigidbody would cause issues when changing scene so it's
safer to just not delete the auto generated rigidbody as it has little
impact on leaving it intact after the object is disabled.