fix(Controller): prevent crash if model has no material colour
If the controller model had no material colour and the controller
opacity or colour is attempted to be changed, then it would crash
because it attempted to find a colour on the material object.

This fix adds a check to see if there is a `_Color` property to
prevent the crash.