fix(Interaction): ensure pointer use activation honours valid hand
Previously, the Pointer Activate Use Action setting would always
activate the use function on an interactable object even if the Valid
Use Hand was set to ignore the hand the pointer was coming from.

This fix ensures the Pointer Use Action honours the Valid Use Hand
setting and if the pointer hand is not valid then teleport function
will also default to working.