fix(Interaction): ensure touch highlight works with multiple renders
Previously, if an object had a renderer and then had children which
also had renders, the main object's renderers would only be used and
the child renderers would be ignored meaning the highlight mechanism
didn't affect all child objects correctly.

This fix ensures that all renderers are dealt with appropriately.