fix(Interaction): improve door gizmos
Improves visualizing the opening direction of doors a lot. There were
quite some situations so far where gizmos would point into the wrong
direction. These should be much fewer now.

In addition when a hinge joint is added manually this is now taken as
the input for the gizmos and the direction is also derived from it
resulting in a much better editor experience.