feat(Controls): add new Panel Menu 2D control
Add Panel Menu control that can attach to VRTK_InteractableObject.
The control is comprised of a prefab and controller scripts.
Together these enable the display of a 2D context menu panel.which
is opened by pressing the assigned controller touchpad vector position
(up/down/left/right). Once a panel is open the controllers pass on
basic interaction events to the panel, these events can be subscribed
to and used to carry out custom UI actions inside the panel.

See the demo scene for a basic working example
[ 040_Controls_Panel_Menu ].