fix(Interaction): use game object instance id as ref for highlighting
Previously, the game object when highlighted would use the name of the
game object to store a reference to it's original colour. However, if
child game objects shared the same name then this would break and cause
a crash with the highlighting.

This fix ensures a unique reference to the game object is used, which
is the instance ID of the game object.