fix(Teleport): add floor height tolerance parameter
The `Floor Height Tolerance` parameter is used to determine how much
of a change between the current floor `y` and the previous floor `y`
has happened before considering it valid to drop to the nearest floor.

This setting can be used to prevent unnecessary teleport events when
the floor below the user is uneven but should not be considered
different enough to force the user to drop to a new location on such
a small variation change in their `y` position.