feat(Climbing): Added Climbing Stamina
Add climbing stamina class that extends the existing VRTK_PlayerClimb
class. The VRTK_PlayerClimbStamina script is added to the CameraRig in
of place the previous climbing script.

The players min reach and max reach can be set where the distance in
the xz plane between the controller and the head determine the effect
on stamina. A distance less than min reach will not decrease stamina
while a distance beyone max reach will apply the maximum damge to
stamina.

The player can regenerate stamina when either no hands are holding a
climbable object, or when both hands are holding climbable objects.
The VRTK currently doesn't support grabbing one object with both hands
but should work when it is implemented.

Example 039_CameraRig_ClimbingStamina has been added to demonstrate
the climbing stamina as well as a useful grip stamina visualization
prefab which has been added.