fix(UIPointer): ensure blocking objects prevent ui ray
The `UI_CONTROL_OFFSET` const was originally implemented to solve an
old issue which has since been resolved in another way but the
original solution using this const was left in causing an issue where
objects that were very close to the canvas (within the threshold)
were allowing the ui ray to pass through the object.

The const is no longer valid or required so removing it fixes the issue
with the ray passing through close objects.