fix(UIPointer): prevent central button flickering on hover
The `UI_CONTROL_OFFSET` const was previously removed as it was thought
it was no longer required. However, the offset does provide a mechanism
for preventing a discrepancy in hit distance that will cause a UI
element if centrally placed in the canvas to constantly hover enter and
immediately hover exit repeatidly. This causes the button flicker as
the button is constantly being hovered over and then out.

The fix is to put the `UI_CONTROL_OFFSET` back in but just set the
offset at a very low value so it stops the ui raycast at a blocking
object but will not trigger the discrepancy with the button flickering.