feat(Locomotion): add option to only touchpad walk on button press
A new option has been added to the Touchpad Walking script that
allows a button to be specified for enabling movement. Only when
that button is being pressed will the touchpad walking actually
move the user.

The Controller Events script has also been updated to allow for this
with a new public method that checks to see if a given button alias
is being pressed.