fix(Teleport): ensure position change over tiny distance change
Previously, if the target position was very small then the lerp time
would be bigger than the elapsed time meaning the lerp would never
have a second pass and therefore the position was never updated.

This fix checks to see if the current position is not equal to the
target position after the lerp time has finished and if they're not
then it force moves the transform position to the target location.