feat(AdaptiveQuality): add Adaptive Quality
Adaptive Quality dynamically changes rendering settings to maintain VR
framerate while maximizing GPU utilization.

It includes a debug visualization and supports both command line
arguments and keyboard shortcuts. A new example scene uses way too many
lights to bring down the GPU to show the framerate getting stabilized
by the script.

Because this is based on The Lab Renderer from Valve their license is
included and mentioned in the files that are derived from it.