feat(Teleport): add play space falling restrictions on height adjust
The play space falling mechanism on the Height Adjust Teleport has
been updated now to allow for a restriction to prevent play space
falling if one/either/both of the controllers are still hovering
over the existing standing floor.

Previously, if the headset looked over a ledge and the raycast
detected a new lower floor, it would immediately teleport the
user down to the new floor. However, now with this new setting
it's possible to lean over the ledge and look down as long as
one of the controllers (or both) is still hovering over the
existing standing floor.