feat(Pointer): add bezier beam height limit based on controller angle
The bezier pointer beam maximum projected height can now be limited
when the controller gets past a certain rotational angle.

Previously, if the controller was rotated to be pointing straight up
then the bezier beam would be projected the full length of the beam
into the sky and then curved back down to the ground. Whilst this
is useful for being able to teleport up to high places, it can also
be detremental in allowing teleporting to high places and also look
unsightly when attempting to teleport somewhere close and hihger up
as the beam would project way up into the air.

The new `Beam Height Limit Angle` can be set between 1 or 100 degrees.