fix(Controller): ensure touchpadTouched is correctly reset to false
The Touchpad Axis Changed code in Controller Events was setting the
touchpad touched state to true every time there was a change in the
touchpad axis. This meant that on release of the touchpad, the axis
would change again and the touchpad touch state would be true even
though the touchpad was no longer being touched.

The fix is to set the correct variable in the `SetButtonEvent` method
so the `touchpadTouched` variable is not incorrectly set.