fix(Controller): prevent error on cache controller tracked object
If the controller isn't ready then SteamVR sets the id to be
the max uint value and if this happens then the cached controller
attempts to add the controller to the array even though it will
already exist. The best solution is to just ignore operations on
caching until the controller is ready (i.e. when the index is a
reasonable number).