fix(Interaction): made knob calculation work in more cases
This is work in progress. The knob calculation is a beast and hard
to get right. It now works in one problematic case but breaks the
scene 25. I have to investigate if the sample knobs are setup
incorrectly or if the calculation is off.

Final text:

The constraints are now enforced through a configurable joint which
gives more options (and also supports moving in the future). The
knobs can also be freely rotated now.