fix(Tooltips): ensure tips are hidden when controller is hidden
Previously, if the controller was hidden the controller tooltips
would stay visible. Due to the new Controller Actions events
it's now possible to know when a controller is being hidden or
shown and these events are now subscribed to by the
Controller Tooltips script and the tooltips are hidden or shown
accordingly.