fix(Interaction): remove limiting max angular velocity on throw
The Throw section of the Interact Grab script would limit the
`maxAngularVelocity` of the rigidbody to the current `angularVelocity`
of the thrown object, this is from the original SteamVR_TestThrow
script and assuming it is there to lock the rotation of the object
to the initial spin, but it would seem that an object should be
able to continue to spin faster if required.

Also, doing this breaks the `Track_Object` grab mechanic as it prevents
it from being rotated at a normal speed once dropped.