fix(Pointer): set destination correctly for constant beam
If the `holdButtonToActivate` variable was set to false so there
was a constant pointer beam and if the pointer toggle and pointer set
alias buttons were different then the destination set event would be
emitted at the wrong time.

It would not emit the destination set event whilst the pointer beam
was still active, but would then allow the destination set event
after the pointer beam had been turned off.

This has been resolved with some additional checking in the pointer
set logic.