feat(Interaction): use new highlighter component on interactable object
The Interactable Object script no longer contains code for highlighting
the object, instead it utilises another Highlighter component that is
added to the game object.

The concept of the Highlighter components is they deal with
highlighting logic and will differ depending on the type of
Highlighter component. If no Highlighter component is added then the
default `Material Color Swap Highlighter` will be used.

There are currently two Highlighters that can be used:

 * `VRTK_MaterialColorSwapHighlighter` works the same way as the
 previous implementation on the Interactable Object script by looping
 through the colors on each material of the object and changing the
 colour. This is fast and simple but breaks Unity Draw Call Batching
 because the material instance is changed.
 * `VRTK_SolidMaterialSwapHighlighter` works by creating a new set
 of materials for the Renderer that contain the highlight colour.
 Then when highlighting happens, the highlight materials are swapped
 with the original materials, which shouldn't break batching as the
 material instances are not changed.