feat(Highlighters): add highlighter composition scripts
A new Highlighter composition concept has been added that allows a
separate component to deal with highlighting of objects.

Previously, the Interactable Object and the Controller highlighting
were done with the same logic in two different scripts and it also
meant that there was only one way to highlight these elements.

The new highlighter scripts mean the logic for highlighting is in
one script but other highlighters can be created and added to an
object to provide alternative highlighting methods.