Create VRTK_TouchpadVelocity.cs
TouchpadVelocity Script

This script is a modified version of the VRTK Touchpad Velocity script.

The function is to provide velocity to the Player Presence rigid body to emulate space suit thruster.

I have included methods to determine which direction the PlayerPresence headset is looking for orientation mapping. This keeps button presses relative to where the player is looking.