fix(Components): simplify getting components and child components
The `GetComponentsInChildren` method will also return any types
found on the current object as well so it's not required to do a
`GetComponents` to affect the current object and then followed by
`GetComponentsInChildren` to affect the child objects.

`GetComponentsInChildren` is enough to deal with all current object
types and all child types as well. It's inefficient to make the
call twice and update the same object in both calls.

There are still occasions where a check to see if the item exists
on the current object using `GetComponent` is valid, because it
it a cheaper call to execute and if the object contains the
component then doing the more expensive `GetComponentInChildren`
is not required to be made.