feat(Highlighters): add outline copy highlighter
A new highlighter has been added that puts an outline around the
object that it is applied to. It achieves this by making a copy of
the object (or a provided mesh) and puts an outline shader on the
object, which is toggled on when the highlight occurs.

The highlighters have also been updated to have an `active`
parameter on them which is used by scripted components to know which
highlighter is the one to use when found on the component (as it may
be required to put multiple highlighter components onto an object
to toggle on and off).

The Controller Actions script has also been updated to allow the
sub elements of the controller to have a custom highlighter provided
for each so it can override the main highlighter if needed.