fix(Interaction): stop crash with interactable object switching scene
If an interactable object was used to switch a scene then the
`OnDisable` method would be called and the `ForceStopInteracting`
method would be set up to run at the end of the frame via co-routine.

The problem was, the scene load destroys the objects and runs the
garbage collector on them meaning that when the co-routine executes
the variables are no longer available to iterate through therefore
causing the crash.

The solution is to check if the variable is valid before proceeding.