feat(InteractGrab): add toggle to autoscale throw velocity with player
A boolean toggle is now used to determine if the user wants the throw
velocity to scale up and down with the camera rig. To get around this
issue before, you had to manually enter the scale in the mulitpier
slot which is error prone and a bit confusing. This toggle keeps all
velocities used to throw in true world space coordinates, so dynamic
size adjustments and throwing just work without any additional input
needed from the developer.