feat(Climbing): climb release can height adjust or physics fall
Instead of assuming a physics based fall is required, the climbing
object now checks if we at or below ground level. If it is, it will
allow the VRTK_HeightAdjustTeleport object to move you up rather
than assuming you must enter a physics falling state.

The ground plane of example 37 (ClimbingFalling) now uses a box
collider on the ground. This should remove the falling through the
floor bug that was reported.