fix(Headset): detect collision end on collider disable and destroy
The `OnTriggerExit` method does not get called if an object is
colliding with the headset and is then disabled or destroyed meaning
the headset collision ended event is never thrown. This means that
if the headset fade is active, it never deactivates.

This fix checks to see if the current collider is still valid in an
Update method and if it's not then it force calls the collision ended
event.