fix(RadialMenu): only change the axis if the touchpad is being touched
SteamVR seems to sent a TouchpadAxisChanged event when the touchpad
is released. This is causing an OnOverEnter event to fire out of the
RadialMenu after the OnHoverExit event is fired. Because of this,
you'll always get an OnHoverEnter event at the end of picking your
finger up off of the touchpad.

Keep track of if the touchpad is currently being touched or not. Only
trigger the axis change if it is during an active touch.