fix(Pointer): ensure custom pointer objects retain set material
Previously, the custom pointer objects were having their materials
overriden by the default pointer material which meant that any custom
objects didn't look correct.

This fix ensures that only the colour and emission colour on the
pointer objects is changed and not the material.

The inspector settings for the pointers has also had a bit of a tidy
up and headers added to make them easier to understand.

The Darken method from the Highlighter has also been moved to a public
static method in Utilities so it can be used in multiple places as it
now is.