chore(Interaction): prevent unhighlight on using action
There is no point in stopping the highlight of an object when using it
because an object can be used without being grabbed. The only time an
object highlight needs turning off is if it is being grabbed, which in
turn will prevent highlighting on use if the item is already being
grabbed.