chore(SDK): move sdk specific code into sub file with only methods
The SDK bridge now is a dumb file that loads in an active SDK
(currently only SteamVR) which is a script that implements an SDK
interface (via abstract clas) to ensure the SDK meets the requirements
for the SDK Bridge to function correctly.

This will enable multiple SDK scripts to co-exist but can use Unity
environment variables to determine which SDK to load in.

The controller element paths have also been put into a method call
rather than a static variable as this will be more generic and
easier to customise for different SDKs.