feat(SDK): allow controller element paths to be hand specific
The SDK `GetControllerElementPath` method now takes an optional `hand`
parameter that can be used to return a different model path if the SDK
uses difference paths for each controller hand model.

For instance, the Oculus Touch controllers are physically different
models for left and right hands, and the paths to the 3D model parts
have different paths.