feat(Highlighter): add new generic Reset method for highlighters
The Base Highlighter now contains a Reset method that is required to
be implemented in any Highlighter class that has the responsibility of
setting up the given highlighter to a default state. This is useful
if an object changes materials during runtime and the highlighter needs
resetting to take this new data into consideration.