feat(Interaction): add multi interactable object gun example
A new gun model has been added that has a working slide and safety
switch along with an animated trigger based on the trigger axis on
the controller.

The new gun can be seen in `021_Controller_GrabbingObjectsWithJoints`
and can be picked up with one hand and then the slide can be grabbed
with the other hand.