feat(Tooltips): add method to update tooltip text at runtime
It's now possible to update a tooltip text at run time by using the
`UpdateText` method on either the Object Tooltip or Controller Tooltip
script.

If any advanced updates are required then the public variables can be
changed at runtime and the `Reset` method can be called to
re-initialise the tooltips.