fix(MonoBehaviour): rename Reset methods to not clash with Unity Reset
The Unity MonoBehaviour has a pre-defined method called `Reset` that is
executed whenever the user adds a component to a game object or clicks
the Reset menu option.

This means that any existing class that had a `Reset` method was
actually tying into the Unity editor Reset calls which is incorrect
as the state should not be reset on these actions.

The solution is to rename the relevant `Reset` methods in the VRTK
scripts to something else so they don't get called when the Unity
`Reset` method is executed.