feat(Highlighter): specify whether highlighter uses cloned object
It's possible that a highlighter will clone the object it is
highlighting to apply the highlight effect to. This is how the
outline highlighter works as the cloned object displays the outline
whilst the original object stays in place.

This commit allows the base highlighter to know and report if the
implemented highlighter is using a cloned object. This can be
useful if another script using highlighting needs to know the
fundamentals of how the highlighter is about to work.