feat(SnapDropZone): add prefab to allow for set drop zones for objects
The new Snap Drop Zone prefab allows an interactable object to be
placed in a specific areas. When the interactable object is held over
a snap drop zone, the destination for the dropped object is highlighted
and when the object is released it's position, scale and rotation are
matched to the highlighted object's settings.

An example scene has been added to show the prefab in use.