feat(UIPointer): add click ui elements with pointer collision
A new setting has been added that will activate the click action on
the UI Pointer if the object the UIPointer is attached to collides
with a valid UI canvas. This allows a UI element to be clicked simply
by activating the UI raycast and touching the UI element.

This works by the ray is active and pointing at a UI element then the
collision forces a click and the UI element that is being touched by
the raycast is considered clicked.