fix(Interaction): prevent button overrides being reset incorrectly
There was an issue where if an interactable object had a grab override
button set and an object was grabbed, the whilst grabbed the controller
no longer touched the object (whilst still grabbing) then the button
overrides would be reset back to the original button.

This was because the StopTouching code would trigger as the controller
and object were no longer touching and the reset buttons code would
execute even if the object was still being grabbed (or used).

It was easy to recreate by grabbing a door that used Rotator_Track
and then moving the controller out of the door collider and then trying
to release the door. It would not release with the grab override
because the grab button had been reset back to the default.

This fix now only resets the override if the object is not being
grabbed or used respectively.