feat(Scaling): add an example to show how to scale an object using one or two hands.
This is in support of issue thestonefox/VRTK#31

Extend VRTK_InteractableObject to demonstrate an approach to uniform object scaling. Objects in the demo scene are configured with "Precision Grab" either on or off. Scale happens at the objects center and can appear to pull away from the controller when using Precision Grab and you are shrinking the object.