feat(UICanvas): add ability to drag and drop UI elements on canvas
Two scripts have been added to provide a customised drag and drop
function for UI elements.

The `VRTK_UIDraggableItem` script is applied to any UI Element inside
a `VRTK_UICanvas` object and will allow for the UI Element to be
dragged around the canvas and dropped onto the canvas in a new
location.

The `VRTK_UIDropZone` script provides a restriction for where the
draggable items can be dropped. If the draggable item is restricted
to a UI Drop Zone then it can only be dropped if it is being hovered
over a drop zone otherwise when it is dropped, it will snap back
to it's original position.

The UI Example scene has been updated with a new canvas to show the
drag and drop ability.