fix(Interaction): ensure track object snap handle uses same orientation
The snap handles on a Track Object grab mechanic would always be
inverted due to the way the code inverted the rotation to ensure the
snap handle could be rotated towards.

This fix simply flips the snap handle rotations on object Awake so
they are already inverted before they get inverted again in the
fixed update.