fix(Interaction): reactivate collisions when an object is ungrabbed
The ability to pause collisions on the object when it is grabbed would
cause an issue if the object was dropped (ungrabbed) whilst the
pause collisions was still active where the object could fall through
the wall as collisions were still disabled on it.

The fix is to unpause the collisions as soon as the object is dropped.