feat(Interaction): extract controller visibility to own script
  > ** Breaking Changes **

The Interactable Object now no longer deals with the controller
visibility state and there is also no global controller visibility
state on touch, grab and use of an interactable object.

The new Interact Controller Appearance script is now applied to any
Interactable Object Game Object that requires the controller to be
toggled visible/hidden when the controller touches/grabs/uses the
object the script is applied to.

This separates out optional logic, ensures no global overrides need
to be handled and cleans up the Interactable Object script.

The example scenes that used the controller visibility settings
have been updated to use the new script.