feat(Interaction): combine secondary grab actions into one parameter
  > ** Breaking Changes **

Previously, the secondary grab action was a drop down with a custom
option allowing to provide a grab action script.

This has now been simplified into one field which is simply the
ability to provide a secondary grab action script. There is also
now a Swap Controller grab action script that provides the same
functionality that the `Swap Controller` option provided in the
previous incarnation of the setup.

Providing no script is the same as the previous `No Action` setting.

If no script is provided and a secondary grab action script is present
on the object, then it will be used at runtime.

The custom Interactable Object editor script has now also been removed
as it is no longer required as the Interactable Object script is much
more simplified with all optional settings now extracted into their
own scripts.