feat(Interaction): add option to apply point distance on throw
Previously, the object velocity on throw was always being upadted to
include the distance of the object from the attach point on the
controller. Whilst this is suitable in some circumstances (e.g. a long
stick with a ball on the end) it is not appropriate in all cases so
it has now been made an option to allow it to be set accordingly.