feat(tag_or_class_string): remove tag_or_class string option
  > ** Breaking Changes **

Previously, a number of scripts contained a string parameter that would
be used to denote a Tag Or Script that should be ignored by the
relevant script. For example, the teleporter script would ignore any
objects that had the given string for a tag or an attached script.

As the Tag or Script Policy List script handles all of this
functionality, it means that the string parameter is largely obsolete
and should be removed to keep the functionality within the Policy
List script.

All of the example scenes have been updated to utilise Policy Lists
now instead of a given script or string name.