feat(Interaction): add highlighting options for IOs and SimplePointer
Adding the new VRTK_Destination Highlighter to the SimplePointer will
use the new options in InteractableObject to highlight the object
when the pointer enters.  When the pointer exits, the original material
will be reapplied.

In VRTK_InteractableObject there is a new panel for Targetting Options
and you can set HighLightTarget to true as well as selecting a
HighLightTargetMaterial.  If either of these are false/null, nothing
will happen.