chore(SDK): add support for a wider range of controller options
  > ** Breaking Changes **

The controller now supports an additional axis for the grip button
and the application menu button has been renamed to `Button One` as
other controllers may utilise this button for other purposes and a
generic name is more appropriate.

The grip and the button one can also register touch events in case
other controllers support touch on these input elements..

The SDK Manager has also been updated to now require the GameObjects
that contain the controller models to be supplied to the SDK Manager
so it can know where the renderers are for each controller and to
access the sub elements of the controller.

A button has been added to the SDK Manager custom inspector that when
clicked will attempt to find the linked objects automatically.

This will help in circumstances where the controller models are not
part of the actual tracked controller GameObject hierarchy.

The Controller Tooltips prefab has also been updated to reflect the
new Button One change.