fix(Pointer): unparent simple pointer holder to fix scaling issue
Previously, the Simple Pointer holder was a child of the controller
object meaning any scaling to the parent camera rig would affect
the scale of the Simple Pointer holder. This meant that if the
camera rig scale was changed during runtime, then the scale of the
Simple Pointer would be wrong.

The Simple Pointer holder is now a top level object meaning it's
scale is not affected by the scaling of any other objects.