fix(Pointer): prevent bezier pointer clipping
The Bezier Pointer beam can occasionally clip through the floor due
to the way the curve is calculated based on 1 forward point and 1
downward point.

This fix casts additional rays along the curve to check for early
collisions and if a collision is found then the target location
is updated to prevent clipping.