fix(Interaction): ensure using object is reset on interact use on stop
The Interact Use variable `Using Object` was not being reset when an
object was no longer being interacted with using a pointer, meaning
that if the pointer stopped the using action then the controller could
not start the using action again because the `Using Object` was still
set.

This fix ensures that whenever `StopUsing` is called on an object that
the controller is also reset to keep it in a clean state.