fix(SDK): catch errors when sdk scripting define symbol is not set
If the `VRTK_SDKManager` script has not been selected in the editor
and the selected SDK has not been able to set the Unity scripting
define symbol, then at runtime a valid SDK would exist but it would
fallback to the Fallback SDK and the relevant scripts that rely
on the SDK properties would fail and throw errors.

This fix catches those errors and provides a helpful message giving
a hint as to where the problem may be for the user.