fix(AdaptiveQuality): remove sdk error when using debug visualisation
If the debug visualisation parameter was ticked and the scene was being
run from the Unity editor, then the `OnValidate` method would run
before the `Awake` method for the SDKManager meaning no SDK was set.

This caused the fallback SDK to be used and therefore throwing the
error about the Fallback SDK being used.

The fix just checks to see if the SDKManager instance has been created
before trying to access something with the Device Finder.